# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.singleTodoResponse:
    properties:
      todo:
        $ref: '#/definitions/handler.todoResponse'
    type: object
...
  handler.todoCreateRequest:
    properties:
      todo:
        properties:
          status:
            type: boolean
          title:
            type: string
        required:
        - status
        - title
        type: object
    type: object
  handler.todoListResponse:
    properties:
      todos:
        items:
          $ref: '#/definitions/handler.todoResponse'
        type: array
    type: object
  handler.todoResponse:
    properties:
      status:
        type: boolean
      title:
        type: string
    type: object
  handler.todoUpdateRequest:
    properties:
      todo:
        properties:
          status:
            type: boolean
          title:
            type: string
        type: object
    type: object
...
  utils.Error:
    properties:
      errors:
        additionalProperties: true
        type: object
    type: object
host: 127.0.0.1:8585
info:
  contact: {}
  description: Conduit API
  title: Conduit API
  version: "1.0"
paths:
...
  /todos:
    get:
      consumes:
      - application/json
      description: Get todo List
      operationId: todos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.todoListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Error'
      security:
      - ApiKeyAuth: []
      summary: Get Todo List
      tags:
      - todo
    post:
      consumes:
      - application/json
      description: Create an todo. Auth is require
      operationId: create-todo
      parameters:
      - description: Todo to create
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/handler.todoCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.singleTodoResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Error'
      security:
      - ApiKeyAuth: []
      summary: Create an todo
      tags:
      - todo
  /todos/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an todo. Auth is required
      operationId: delete-todo
      parameters:
      - description: id of the todo to delete
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete an todo
      tags:
      - todo
    get:
      consumes:
      - application/json
      description: Get todo by id
      operationId: get-todo
      parameters:
      - description: id of the todo to get
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.todoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Error'
      security:
      - ApiKeyAuth: []
      summary: Get Todo By Id
      tags:
      - todo
    put:
      consumes:
      - application/json
      description: Update an todo. Auth is required
      operationId: update-todo
      parameters:
      - description: id of the todo to update
        in: path
        name: id
        required: true
        type: string
      - description: Todo to update
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/handler.todoUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.singleTodoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Error'
      security:
      - ApiKeyAuth: []
      summary: Update an todo
      tags:
      - todo
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"